refactor: 날짜 문자열 처리 유틸함수 통합
