+++ src/utils/dateToString.ts
@@ -8,3 +8,15 @@ export const dateToString = (dateObject: Date) => {
 export const stringDateIsToday = (stringDate: string) => {
   return stringDate === dateToString(new Date());
 };
+
+export const stringDateMove = (stringDate: string, dx: number) => {
+  const [y, m, d] = stringDate.split('-').map((e) => Number.parseInt(e));
+  const dateObj = new Date(y, m - 1, d + dx);
+  return dateToString(dateObj);
+};
+
+export const stringDateMoveKor = (string: string, dx: number) => {
+  const [y, m, d] = string.split('-').map((e) => Number.parseInt(e));
+  const date = new Date(y, m - 1, d + dx);
+  return `${date.getMonth() + 1}월 ${date.getDate()}일`;
+};
